+++ app/src/main/java/com/blackboxembedded/WunderLINQ/comms/BLE/BLEbus.java
@@ -15,12 +15,11 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-package com.blackboxembedded.WunderLINQ.protocols;
+package com.blackboxembedded.WunderLINQ.comms.BLE;
 
 import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
 
-import com.blackboxembedded.WunderLINQ.comms.BLE.BluetoothLeService;
 import com.blackboxembedded.WunderLINQ.hardware.WLQ.Data;
 import com.blackboxembedded.WunderLINQ.hardware.WLQ.Faults;
 import com.blackboxembedded.WunderLINQ.MyApplication;
