+++ app/src/main/java/com/blackboxembedded/WunderLINQ/comms/BLE/KeyboardHID.java
@@ -15,7 +15,7 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-package com.blackboxembedded.WunderLINQ.protocols;
+package com.blackboxembedded.WunderLINQ.comms.BLE;
 
 import com.blackboxembedded.WunderLINQ.MyApplication;
 import com.blackboxembedded.WunderLINQ.R;
