+++ app/src/main/java/com/blackboxembedded/WunderLINQ/hardware/WLQ/WLQ_N.java
@@ -21,7 +21,7 @@ import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
 import android.util.Log;
 
-import com.blackboxembedded.WunderLINQ.protocols.KeyboardHID;
+import com.blackboxembedded.WunderLINQ.comms.BLE.KeyboardHID;
 import com.blackboxembedded.WunderLINQ.MyApplication;
 import com.blackboxembedded.WunderLINQ.R;
 import com.blackboxembedded.WunderLINQ.Utils.Utils;
