+++ app/src/main/java/com/blackboxembedded/WunderLINQ/hardware/WLQ/WLQ_X.java
@@ -24,7 +24,7 @@ import android.util.Log;
 import com.blackboxembedded.WunderLINQ.MyApplication;
 import com.blackboxembedded.WunderLINQ.R;
 import com.blackboxembedded.WunderLINQ.Utils.Utils;
-import com.blackboxembedded.WunderLINQ.protocols.KeyboardHID;
+import com.blackboxembedded.WunderLINQ.comms.BLE.KeyboardHID;
 
 import java.util.Arrays;
